let the formants simply be ints
We will need the negative values anyway
once we start computing covariance.

+++ point.c
@@ -6,10 +6,10 @@
 
 #include "point.h"
 
-static uint16_t
+static int
 getnum(char** line)
 {
-	uint16_t F = 0;
+	int F = 0;
 	char* p = NULL;
 	const char* e = NULL;
 	while (*line && isspace(**line))
